% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_combo.R
\name{calendar_combo_extract}
\alias{calendar_combo_extract}
\title{Extract combination of events (pattern or part of it)}
\usage{
calendar_combo_extract(calendar, event1, n_months, event2,
  output = "pattern")
}
\arguments{
\item{calendar}{string(s) of calendar, months in original order (i.e. recent -> older)}

\item{event1}{code(s) of the earlier/older event(s) of interest}

\item{n_months}{number of months between events (numerical)}

\item{event2}{code(s) of the later/recent event(s) of interest}

\item{output}{chose if return "event1", "event2" or the whole "pattern" detected (default: "pattern")}
}
\value{
the combination/pattern of interest
}
\description{
Extract combination of events (pattern or part of it)
}
\examples{
dat_calendar <- c("333000BPPP", "000000BPPP", "32100000000BPP", "56700ABC")
calendar_combo_extract(calendar = dat_calendar,
event1 = "B", n_months = 6, event2 = c(1:9))
calendar_combo_extract(calendar = dat_calendar,
event1 = "B", n_months = 6, event2 = c(1:9),
output = "event2")
}
