% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_recode.R
\name{calendar_recode}
\alias{calendar_recode}
\title{Recode Data - DHS Calendar}
\usage{
calendar_recode(x, recodebook = DHScalendR::dhs_calendar_codebook,
  code = "Recode", recoding = "Meaning")
}
\arguments{
\item{x}{is a vector of encoded data}

\item{recodebook}{is a data frame containings information about code}

\item{code}{name column of recodebook with the code of x (character)}

\item{recoding}{name column of recodebook with meaning of the code (character)}
}
\value{
recoded vector x
}
\description{
Recode Data - DHS Calendar
}
\examples{
events <- c("0","0","P","P","T","A","A", NA, "X", "3","3")
calendar_recode(events)

}
