% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_split.R
\name{calendar_split}
\alias{calendar_split}
\title{Split Calendar Into Individual Months}
\usage{
calendar_split(dhs, interview = "V008", calendar = "VCAL.1")
}
\arguments{
\item{dhs}{dataframe containing the following variables}

\item{interview}{name as string of the "Date of interview (CMC)" variable}

\item{calendar}{name as string of the "Calendar" variable}
}
\value{
same dataframe + calendar divided with one column "CMCx" per month
}
\description{
Split Calendar Into Individual Months
}
\examples{
dataset <- dhs_extract[95:100, c("V008", "V025", "VCAL.1")]
dataset_plus_CMCx <- calendar_split(dataset, "V008", "VCAL.1")
dataset_plus_CMCx[ ,1:10]
}
