% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_subset.R
\name{calendar_subset}
\alias{calendar_subset}
\title{Subset Months from DHS Calendar}
\usage{
calendar_subset(calendar, interview, starting, ending)
}
\arguments{
\item{calendar}{string(s) of calendar, months in original order (i.e. recent -> older)}

\item{interview}{date of interview (CMC format, numeric)}

\item{starting}{date of first month of interest (CMC format, numeric)}

\item{ending}{date of last month of interest (CMC format, numeric)}
}
\value{
a vector containing the strings of months of interest
}
\description{
Subset Months from DHS Calendar
}
\examples{
df <- DHScalendR::dhs_extract[5:9, ]
calendar_subset(calendar = df$VCAL.1, interview = df$V008, starting = 1379, ending = 1385)
}
