% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc.R
\name{cmc_from_Date}
\alias{cmc_from_Date}
\title{Transform Date to CMC Format}
\usage{
cmc_from_Date(x)
}
\arguments{
\item{x}{is a date format object}
}
\value{
date in CMC format
}
\description{
Transform Date to CMC Format
}
\examples{
cmc_from_Date(as.Date("1963-05-13")) #761

}
