% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc.R
\name{cmc_from_Year}
\alias{cmc_from_Year}
\title{Transform Years to CMC Format}
\usage{
cmc_from_Year(x)
}
\arguments{
\item{x}{is a year with decimal}
}
\value{
date in a CMC format
}
\description{
Transform Years to CMC Format
}
\examples{
#Will run warning the first time
cmc_from_Year(1989.296) #1072

}
