% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc.R
\name{cmc_to_Date}
\alias{cmc_to_Date}
\title{Transform CMC Format to Date Format}
\usage{
cmc_to_Date(x)
}
\arguments{
\item{x}{is a numerical CMC date format}
}
\value{
date in Date format
}
\description{
Transform CMC Format to Date Format
}
\examples{
cmc_to_Date(761) #1963-05-01

}
