% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc.R
\name{cmc_to_Year}
\alias{cmc_to_Year}
\title{Transform CMC Format to Years}
\usage{
cmc_to_Year(x)
}
\arguments{
\item{x}{is a numerical CMC date format}
}
\value{
a year with decimal
}
\description{
Transform CMC Format to Years
}
\examples{
cmc_to_Year(1072) #1989.25

}
