% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_combo.R
\name{event_combo}
\alias{event_combo}
\title{Check if Combination of Events in calendar}
\usage{
event_combo(calendar, event1, n_months, event2)
}
\arguments{
\item{calendar}{string(s) of DHS calendar, months in original order (i.e. recent -> older)}

\item{event1}{code(s) of the earlier/older event(s) of interest}

\item{n_months}{number of months between events (numerical)}

\item{event2}{code(s) of the later/recent event(s) of interest}
}
\value{
wheter or not each calendar items contain the combination of events
}
\description{
Check if Combination of Events in calendar
}
\examples{
dat_calendar <- c("3300000BPP", "5000TPP111", "0123005670")
event_combo(calendar = dat_calendar, event1 = "P", n_months = 6, event2 = c(1:9))
event_combo(calendar = dat_calendar, event1 = "P", n_months = 4, event2 = c(1:9))
}
