% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_locate.R
\name{event_locate}
\alias{event_locate}
\title{Get location event in a calendar variable}
\usage{
event_locate(
  calendar,
  event,
  sequences = TRUE,
  recent_oldest = "recent",
  start_end = "start",
  interview = 0
)
}
\arguments{
\item{calendar}{string(s) of DHS calendar, months in original order (i.e. recent -> older)}

\item{event}{code(s) of event of interest (can take regex)}

\item{sequences}{look for sequence of event - if FALSE individual events [default: TRUE]}

\item{recent_oldest}{return location most recent or the oldest event in the calendar [default: "recent"]}

\item{start_end}{return location of the start or the end of the event [default: "start"]}

\item{interview}{variable of the date of interview [default: 0]}
}
\value{

}
\description{
Get location event in a calendar variable
}
\examples{
dat_calendar <- c("3300000BPP", "5000TPP111", "0123005670")
event_locate(dat_calendar, "P")
event_locate(dat_calendar, "P", sequences = FALSE)
event_locate(dat_calendar, 0)
event_locate(dat_calendar, 0, recent_oldest = "oldest")
event_locate(dat_calendar, 0, recent_oldest = "oldest", start_end = "end")
event_locate(dat_calendar, 0, interview = c(0, 10, 100))
}
